requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0

# MCP Server dependencies
mcp==1.0.0
psycopg2-binary==2.9.9
PyYAML==6.0.1